Main Page uses Stacked Widget, move graph page to own class
The Main Page now uses a QStackedWidget, with the only page
currently being the graph page.  The existing main page has been
moved to a class GraphPage that contains the plotting widget
and the graph selector.

Move the filter bar to a splitter that is to the left of the stacked widget,
giving a much cleaner look.